% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{loadData}
\alias{loadData}
\title{Load data from database function}
\usage{
loadData(tablename)
}
\arguments{
\item{tablename}{Name of the table in database}
}
\value{
data frame object
}
\description{
This function allows to read data from database
}
\keyword{database}

